backend/requirements.txt
# FastAPI и зависимости
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
celery==5.3.6

# Elasticsearch
elasticsearch==8.12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1

# Google OAuth2
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# HTTP клиент для парсинга
aiohttp==3.9.3
beautifulsoup4==4.12.3
lxml==5.1.0
brotli==1.1.0

# Playwright для браузерной автоматизации
playwright==1.40.0

# OpenAI для категоризации новин
openai==1.55.3

# Push notifications
exponent-server-sdk==2.0.0

# Валидация
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Utilities
python-dateutil==2.8.2

